#346 refactor to simplfy brick module
