+++ src/f09_brick/examples/examples_pandas.py
