+++ src/f09_brick/test_brick_config/test_x_dir_func_.py
@@ -7,7 +7,7 @@ from src.f04_gift.atom_config import (
     type_AcctID_str,
 )
 from src.f08_filter.filter import bridgeunit_shop, filterunit_shop
-from src.f09_brick.dir_func import (
+from src.f09_brick.pandas_tool import (
     filter_single_column_dataframe,
     filter_all_columns_dataframe,
     get_dataframe_filterable_columns,
